LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-animation-length-expected.txt


PASS Animating a custom property of type <length>
PASS Animating a custom property of type <length> with a single keyframe
FAIL Animating a custom property of type <length> with additivity assert_equals: expected "350px" but got "250px"
FAIL Animating a custom property of type <length> with a single keyframe and additivity assert_equals: expected "250px" but got "200px"
FAIL Animating a custom property of type <length> with iterationComposite assert_equals: expected "250px" but got "50px"
